Despawn pipe items after 20 minutes.

+++ common/src/main/java/jagm/classicpipes/blockentity/GoldenPipeEntity.java
@@ -12,12 +12,12 @@ public class GoldenPipeEntity extends RoundRobinPipeEntity {
     }
 
     @Override
-    public int getTargetSpeed() {
+    public short getTargetSpeed() {
         return ItemInPipe.DEFAULT_SPEED * 8;
     }
 
     @Override
-    public int getAcceleration() {
+    public short getAcceleration() {
         return ItemInPipe.DEFAULT_ACCELERATION * 16;
     }
