+++ common/src/main/java/jagm/classicpipes/blockentity/IronPipeEntity.java
@@ -24,12 +24,12 @@ public class IronPipeEntity extends AbstractPipeEntity {
     }
 
     @Override
-    public int getTargetSpeed() {
+    public short getTargetSpeed() {
         return ItemInPipe.DEFAULT_SPEED;
     }
 
     @Override
-    public int getAcceleration() {
+    public short getAcceleration() {
         return ItemInPipe.DEFAULT_ACCELERATION;
     }
